* BlamLib: Copypasta'd a bunch of KSoft.NET classes that are needed to support the new string id implementation (also from KSoft.NET) * BlamLib: Finished implementing the new string id system * BlamLib: Half tested Halo4 support. Right now, I don't really care if it doesn't all work. Vertex definitions will have to wait as well.

+++ BlamLib/BlamLib/TagInterface/Exceptions.cs
@@ -70,7 +70,7 @@ namespace BlamLib.TagInterface
 				Exception inner, 
 				uint offset_field, uint offset_data, 
 				IO.ITagStream tag,
-				byte length, string value) : base(inner, "Invalid String Id")
+				int length, string value) : base(inner, "Invalid String Id")
 			{
 				Debug.LogFile.WriteLine(
 					"Invalid string_id: field@{0:X8}\tdata@{1:X8}\tlength:{2:X2}\tvalue:{3}" + Program.NewLine +
